refactor: move Navigation to common components folder
