+++ src/components/base/index.ts
@@ -4,7 +4,6 @@ export * from './CanvasPlayer';
 export * from './Container';
 export * from './Flex';
 export * from './Hidden';
-export * from './Navigation';
 export * from './Logo';
 export * from './RoundCursor';
 export * from './SEO';
