+++ src/components/common/Navigation/NormalNavigation/ProductPanel/QuestionIcon.tsx
