+++ src/components/common/Navigation/NormalNavigation/ServicesPanel/ServicesPanel.tsx
@@ -3,9 +3,9 @@ import { FC, ReactNode } from 'react';
 import { useTheme } from '@emotion/react';
 
 import { useHover } from '../../../../../hooks';
-import { Box } from '../../../Box';
-import { Container } from '../../../Container';
-import { Flex } from '../../../Flex';
+import { Box } from '../../../../base/Box';
+import { Container } from '../../../../base/Container';
+import { Flex } from '../../../../base/Flex';
 import { NavigationDataItem } from '../../useNavigationData';
 import { useNormalNavigationHeight } from '../useNormalNavigationHeight';
